Perfectly organized worker.py and other files
